Passing safe index by size_t, not by signed char.

+++ demos/spectre_v1.cc
@@ -88,7 +88,7 @@ static char leak_byte(const char *data, size_t offset) {
     }
 
     std::pair<bool, char> result =
-        sidechannel.RecomputeScores(data[safe_offset]);
+        sidechannel.RecomputeScores(static_cast<size_t>(data[safe_offset]));
     if (result.first) {
       return result.second;
     }
